Rename stream_listener_widget.py to stream_event_manager.py
