+++ brainframe_qt/ui/resources/video_items/streams/stream_event_manager.py
